Added syncWritePosition and syncWritePositionAndVelocity. Need to be tested
Former-commit-id: 466cdff5ac424c1f389066870c101718a41ea84a [formerly 55f2922d52cf13b9ccb0086e68efc06618ff9725]
Former-commit-id: 887a1ec2f671b0d6170b0d3ace4bb976bba5d7ab

+++ Robot/Drivers/Dynamixel/MX28/MX28.h
@@ -38,7 +38,7 @@ void MX28_SetResolutionDivider(Dynamixel_HandleTypeDef* hdynamixel, uint8_t divi
 void MX28_SetDGain(Dynamixel_HandleTypeDef* hdynamixel, uint8_t DGain);
 void MX28_SetIGain(Dynamixel_HandleTypeDef* hdynamixel, uint8_t IGain);
 void MX28_SetPGain(Dynamixel_HandleTypeDef* hdynamixel, uint8_t PGain);
-void MX28_SetGoalAcceleration(Dynamixel_HandleTypeDef* hdynamixel, double goalAcceleration);
+void MX28_SetGoalAcceleration(Dynamixel_HandleTypeDef* hdynamixel, float goalAcceleration);
 
 // Interfaces for previously-defined functions
 void MX28_EnterMultiTurnMode(Dynamixel_HandleTypeDef* hdynamixel);
